Fixed sni unit test
